+++ kleiner-brauhelfer/dialogs/dlgbewertungen.cpp
@@ -1,108 +1,130 @@
-#include "tabbewertung.h"
-#include "ui_tabbewertung.h"
-#include <QScrollBar>
+#include "dlgbewertungen.h"
+#include "ui_dlgbewertungen.h"
 #include "brauhelfer.h"
-#include "widgets/wdgbewertung.h"
+#include "settings.h"
+#include "model/textdelegate.h"
+#include "model/datedelegate.h"
+#include "model/sudnamedelegate.h"
+#include "model/spinboxdelegate.h"
+#include "model/ratingdelegate.h"
+#include "model/datetimedelegate.h"
 
 extern Brauhelfer* bh;
+extern Settings* gSettings;
 
-TabBewertung::TabBewertung(QWidget *parent) :
-    TabAbstract(parent),
-    ui(new Ui::TabBewertung),
-    mIndex(-1)
+DlgBewertungen* DlgBewertungen::Dialog = nullptr;
+
+DlgBewertungen::DlgBewertungen(QWidget *parent) :
+    DlgAbstract(staticMetaObject.className(), parent, Qt::WindowMaximizeButtonHint | Qt::WindowCloseButtonHint),
+    ui(new Ui::DlgBewertungen),
+    mSudId(-1),
+    mRowBewertung(-1)
 {
     ui->setupUi(this);
-    connect(bh, &Brauhelfer::discarded, this, &TabBewertung::updateValues);
-    connect(bh->sud(), &SudObject::loadedChanged, this, &TabBewertung::sudLoaded);
-    connect(bh->sud(), &SudObject::dataChanged, this, &TabBewertung::sudDataChanged);
-    connect(bh->sud()->modelBewertungen(), &ProxyModel::layoutChanged, this, &TabBewertung::modelModified);
-    connect(bh->sud()->modelBewertungen(), &ProxyModel::rowsInserted, this, &TabBewertung::modelModified);
-    connect(bh->sud()->modelBewertungen(), &ProxyModel::rowsRemoved, this, &TabBewertung::modelModified);
-    connect(bh->sud()->modelBewertungen(), &ProxyModel::modified, this, &TabBewertung::updateValues);
-    connect(ui->wdgRating, &WdgRating::clicked, this, &TabBewertung::sterneChanged);
+
+    TableView *table = ui->table;
+    ProxyModelSud *proxyModel = new ProxyModelSud(this);
+    proxyModel->setSourceModel(bh->modelSud());
+    proxyModel->setFilterStatus(ProxyModelSud::Abgefuellt | ProxyModelSud::Verbraucht);
+    table->setModel(proxyModel);
+    table->appendCol({ModelSud::ColSudname, true, false, -1, new SudNameDelegate(table)});
+    table->appendCol({ModelSud::ColSudnummer, false, true, 80, new SpinBoxDelegate(table)});
+    table->appendCol({ModelSud::ColKategorie, false, true, 100, new TextDelegate(false, Qt::AlignCenter, table)});
+    table->appendCol({ModelSud::ColAnlage, false, true, 100, new TextDelegate(false, Qt::AlignCenter, table)});
+    table->appendCol({ModelSud::ColBraudatum, false, true, 100, new DateDelegate(false, true, table)});
+    table->appendCol({ModelSud::ColBewertungMittel, true, true, 80, new RatingDelegate(table)});
+    table->build();
+    table->setDefaultContextMenu(false, false);
+
+    table = ui->tableBewertungen;
+    ProxyModel* proxyModelBewertungen = new ProxyModel(this);
+    proxyModelBewertungen->setSourceModel(bh->modelBewertungen());
+    proxyModelBewertungen->setFilterKeyColumn(ModelBewertungen::ColSudID);
+    proxyModelBewertungen->setFilterRegularExpression(QRegularExpression(QStringLiteral("^%1$").arg(mSudId)));
+    table->setModel(proxyModelBewertungen);
+    table->appendCol({ModelBewertungen::ColDatum, true, false, -1, new DateTimeDelegate(false, true, table)});
+    table->appendCol({ModelBewertungen::ColWoche, true, true, 80, new TextDelegate(true, Qt::AlignHCenter | Qt::AlignVCenter, table)});
+    table->appendCol({ModelBewertungen::ColSterne, true, false, 100, new RatingDelegate(table)});
+    table->build();
+    table->setDefaultContextMenu(false, false);
+
+    ui->splitter->setStretchFactor(0, 1);
+    ui->splitter->setStretchFactor(1, 1);
+    ui->splitter->setStretchFactor(2, 4);
+    ui->splitter->setSizes({400, 300, 400});
+
+    connect(ui->table->selectionModel(), &QItemSelectionModel::selectionChanged, this, &DlgBewertungen::onTableSelectionChanged);
+    connect(ui->tableBewertungen->selectionModel(), &QItemSelectionModel::selectionChanged, this, &DlgBewertungen::onTableBewertungenSelectionChanged);
+
+    connect(bh, &Brauhelfer::discarded, this, &DlgBewertungen::updateValues);
+    connect(bh, &Brauhelfer::modified, this, &DlgBewertungen::updateValues);
+    connect(ui->wdgRating, &WdgRating::clicked, this, &DlgBewertungen::sterneChanged);
     connect(ui->wdgBemerkung, &WdgBemerkung::changed, this, [this](const QString& html){setData(ModelBewertungen::ColBemerkung, html);});
+
+    connect(proxyModel, &ProxyModel::layoutChanged, this, &DlgBewertungen::onFilterChanged);
+    connect(proxyModel, &ProxyModel::rowsInserted, this, &DlgBewertungen::onFilterChanged);
+    connect(proxyModel, &ProxyModel::rowsRemoved, this, &DlgBewertungen::onFilterChanged);
+
+    if (bh->sud()->isLoaded())
+        ui->table->selectRow(proxyModel->getRowWithValue(ModelSud::ColID, bh->sud()->id()));
+    else
+        ui->table->selectRow(0);
+    updateValues();
 }
 
-TabBewertung::~TabBewertung()
+DlgBewertungen::~DlgBewertungen()
 {
     delete ui;
 }
 
-void TabBewertung::sudLoaded()
+void DlgBewertungen::saveSettings()
 {
-    mIndex = -1;
-    if (bh->sud()->modelBewertungen()->rowCount() > 0)
-        clicked(0);
-    else
-       updateValues();
+    ProxyModelSud *proxyModel = static_cast<ProxyModelSud*>(ui->table->model());
+    gSettings->beginGroup(staticMetaObject.className());
+    proxyModel->saveSetting(gSettings);
+    gSettings->setValue("splitterState", ui->splitter->saveState());
+    gSettings->setValue("tableState", ui->table->horizontalHeader()->saveState());
+    gSettings->setValue("tableBewertungenState", ui->tableBewertungen->horizontalHeader()->saveState());
+    gSettings->endGroup();
 }
 
-void TabBewertung::sudDataChanged(const QModelIndex& index)
+void DlgBewertungen::loadSettings()
 {
-    if (index.column() == ModelSud::ColStatus)
-    {
-        updateValues();
-    }
+    ProxyModelSud *proxyModel = static_cast<ProxyModelSud*>(ui->table->model());
+    gSettings->beginGroup(staticMetaObject.className());
+    proxyModel->loadSettings(gSettings);
+    proxyModel->setFilterStatus(ProxyModelSud::Abgefuellt | ProxyModelSud::Verbraucht);
+    ui->splitter->restoreState(gSettings->value("splitterState").toByteArray());
+    ui->table->restoreState(gSettings->value("tableState").toByteArray());
+    ui->tableBewertungen->restoreState(gSettings->value("tableBewertungenState").toByteArray());
+    gSettings->endGroup();
+    ui->btnFilter->setModel(proxyModel, FilterButtonSud::Braudatum | FilterButtonSud::Kategorie | FilterButtonSud::Anlage);
 }
 
-void TabBewertung::modelModified()
+void DlgBewertungen::restoreView()
 {
-    const int nModel = bh->sud()->modelBewertungen()->rowCount();
-    int nLayout = ui->layoutBewertungen->count();
-    while (nLayout < nModel)
-    {
-        WdgBewertung *wdg = new WdgBewertung(nLayout++);
-        connect(wdg, &WdgBewertung::clicked, this, &TabBewertung::clicked);
-        ui->layoutBewertungen->addWidget(wdg);
-    }
-    while (ui->layoutBewertungen->count() != nModel)
-        delete ui->layoutBewertungen->itemAt(ui->layoutBewertungen->count() - 1)->widget();
-    for (int i = 0; i < ui->layoutBewertungen->count(); ++i)
-        static_cast<WdgBewertung*>(ui->layoutBewertungen->itemAt(i)->widget())->updateValues();
-    updateValues();
+    DlgAbstract::restoreView(staticMetaObject.className(), Dialog);
+    gSettings->beginGroup(staticMetaObject.className());
+    gSettings->remove("splitterState");
+    gSettings->endGroup();
 }
 
-void TabBewertung::clicked(int index)
+QVariant DlgBewertungen::data(int col) const
 {
-    mIndex = index;
-    for (int i = 0; i < ui->layoutBewertungen->count(); ++i)
-    {
-        WdgBewertung *wdg = static_cast<WdgBewertung*>(ui->layoutBewertungen->itemAt(i)->widget());
-        wdg->setActive(i == mIndex);
-        wdg->updateValues();
-    }
-    updateValues();
+    return bh->modelBewertungen()->data(mRowBewertung, col);
 }
 
-void TabBewertung::on_btnNeueBewertung_clicked()
+bool DlgBewertungen::setData(int col, const QVariant &value)
 {
-    QMap<int, QVariant> values({{ModelBewertungen::ColSudID, bh->sud()->id()}});
-    int index = bh->sud()->modelBewertungen()->append(values);
-    ui->scrollAreaBewertungen->verticalScrollBar()->setValue(ui->scrollAreaBewertungen->verticalScrollBar()->maximum());
-    clicked(index);
+    return bh->modelBewertungen()->setData(mRowBewertung, col, value);
 }
 
-QVariant TabBewertung::data(int col) const
+void DlgBewertungen::updateValues()
 {
-    return bh->sud()->modelBewertungen()->data(mIndex, col);
-}
+    ui->scrollAreaWidgetContents->setEnabled(mSudId >= 0 && mRowBewertung >= 0 && mRowBewertung < bh->modelBewertungen()->rowCount());
 
-bool TabBewertung::setData(int col, const QVariant &value)
-{
-    return bh->sud()->modelBewertungen()->setData(mIndex, col, value);
-}
-
-void TabBewertung::updateValues()
-{
-    Brauhelfer::SudStatus status = static_cast<Brauhelfer::SudStatus>(bh->sud()->getStatus());
-    ui->btnNeueBewertung->setVisible(status == Brauhelfer::SudStatus::Abgefuellt);
-    ui->lineNeueBewertung->setVisible(status == Brauhelfer::SudStatus::Abgefuellt);
-
-    ui->scrollAreaWidgetContents->setEnabled(mIndex >= 0 && mIndex < bh->sud()->modelBewertungen()->rowCount());
-
+    ui->lblWocheBewertungen->setText(data(ModelBewertungen::ColWoche).toString());
     ui->wdgRating->setSterne(data(ModelBewertungen::ColSterne).toInt());
-    ui->tbDatum->setMinimumDate(bh->sud()->getAbfuelldatum().date());
-    ui->tbDatum->setMaximumDate(QDate::currentDate());
     if (!ui->tbDatum->hasFocus())
         ui->tbDatum->setDate(data(ModelBewertungen::ColDatum).toDate());
     ui->wdgBemerkung->setHtml(data(ModelBewertungen::ColBemerkung).toString());
@@ -227,23 +249,23 @@ void TabBewertung::updateValues()
         ui->tbGesamteindruck->setPlainText(data(ModelBewertungen::ColGesamteindruckBemerkung).toString());
 }
 
-void TabBewertung::sterneChanged(int sterne)
+void DlgBewertungen::sterneChanged(int sterne)
 {
     setData(ModelBewertungen::ColSterne, sterne);
 }
 
-void TabBewertung::on_tbDatum_dateChanged(const QDate &date)
+void DlgBewertungen::on_tbDatum_dateChanged(const QDate &date)
 {
     if (ui->tbDatum->hasFocus())
         setData(ModelBewertungen::ColDatum, date);
 }
 
-void TabBewertung::on_btnDatumHeute_clicked()
+void DlgBewertungen::on_btnDatumHeute_clicked()
 {
     setData(ModelBewertungen::ColDatum, QDate::currentDate());
 }
 
-void TabBewertung::setFarbe()
+void DlgBewertungen::setFarbe()
 {
     int value = 0;
     if (ui->rbFarbe_0->isChecked())
@@ -271,78 +293,78 @@ void TabBewertung::setFarbe()
     setData(ModelBewertungen::ColFarbe, value);
 }
 
-void TabBewertung::on_rbFarbe_0_clicked()
+void DlgBewertungen::on_rbFarbe_0_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_1_clicked()
+void DlgBewertungen::on_rbFarbe_1_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_2_clicked()
+void DlgBewertungen::on_rbFarbe_2_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_3_clicked()
+void DlgBewertungen::on_rbFarbe_3_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_4_clicked()
+void DlgBewertungen::on_rbFarbe_4_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_5_clicked()
+void DlgBewertungen::on_rbFarbe_5_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_6_clicked()
+void DlgBewertungen::on_rbFarbe_6_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_7_clicked()
+void DlgBewertungen::on_rbFarbe_7_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_8_clicked()
+void DlgBewertungen::on_rbFarbe_8_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_9_clicked()
+void DlgBewertungen::on_rbFarbe_9_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_10_clicked()
+void DlgBewertungen::on_rbFarbe_10_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_ka_clicked()
+void DlgBewertungen::on_rbFarbe_ka_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_rbFarbe_ka2_clicked()
+void DlgBewertungen::on_rbFarbe_ka2_clicked()
 {
     setFarbe();
 }
 
-void TabBewertung::on_tbFarbe_textChanged()
+void DlgBewertungen::on_tbFarbe_textChanged()
 {
     if (ui->tbFarbe->hasFocus())
         setData(ModelBewertungen::ColFarbeBemerkung, ui->tbFarbe->toPlainText());
 }
 
-void TabBewertung::setSchaum()
+void DlgBewertungen::setSchaum()
 {
     int value = 0;
     if (ui->rbSchaum_0->isChecked())
@@ -370,88 +392,88 @@ void TabBewertung::setSchaum()
     setData(ModelBewertungen::ColSchaum, value);
 }
 
-void TabBewertung::on_rbSchaum_0_clicked()
+void DlgBewertungen::on_rbSchaum_0_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_1_clicked()
+void DlgBewertungen::on_rbSchaum_1_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_2_clicked()
+void DlgBewertungen::on_rbSchaum_2_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_3_clicked()
+void DlgBewertungen::on_rbSchaum_3_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_4_clicked()
+void DlgBewertungen::on_rbSchaum_4_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_5_clicked()
+void DlgBewertungen::on_rbSchaum_5_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_6_clicked()
+void DlgBewertungen::on_rbSchaum_6_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_7_clicked()
+void DlgBewertungen::on_rbSchaum_7_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_8_clicked()
+void DlgBewertungen::on_rbSchaum_8_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_9_clicked()
+void DlgBewertungen::on_rbSchaum_9_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_10_clicked()
+void DlgBewertungen::on_rbSchaum_10_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_ka_clicked()
+void DlgBewertungen::on_rbSchaum_ka_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_ka2_clicked()
+void DlgBewertungen::on_rbSchaum_ka2_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_ka3_clicked()
+void DlgBewertungen::on_rbSchaum_ka3_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_rbSchaum_ka4_clicked()
+void DlgBewertungen::on_rbSchaum_ka4_clicked()
 {
     setSchaum();
 }
 
-void TabBewertung::on_tbSchaum_textChanged()
+void DlgBewertungen::on_tbSchaum_textChanged()
 {
     if (ui->tbSchaum->hasFocus())
         setData(ModelBewertungen::ColSchaumBemerkung, ui->tbSchaum->toPlainText());
 }
 
-void TabBewertung::setGeruch()
+void DlgBewertungen::setGeruch()
 {
     int value = 0;
     if (ui->rbGeruch_0->isChecked())
@@ -481,73 +503,73 @@ void TabBewertung::setGeruch()
     setData(ModelBewertungen::ColGeruch, value);
 }
 
-void TabBewertung::on_rbGeruch_0_clicked()
+void DlgBewertungen::on_rbGeruch_0_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_1_clicked()
+void DlgBewertungen::on_rbGeruch_1_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_2_clicked()
+void DlgBewertungen::on_rbGeruch_2_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_3_clicked()
+void DlgBewertungen::on_rbGeruch_3_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_4_clicked()
+void DlgBewertungen::on_rbGeruch_4_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_5_clicked()
+void DlgBewertungen::on_rbGeruch_5_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_6_clicked()
+void DlgBewertungen::on_rbGeruch_6_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_7_clicked()
+void DlgBewertungen::on_rbGeruch_7_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_8_clicked()
+void DlgBewertungen::on_rbGeruch_8_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_9_clicked()
+void DlgBewertungen::on_rbGeruch_9_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_10_clicked()
+void DlgBewertungen::on_rbGeruch_10_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_rbGeruch_11_clicked()
+void DlgBewertungen::on_rbGeruch_11_clicked()
 {
     setGeruch();
 }
 
-void TabBewertung::on_tbGeruch_textChanged()
+void DlgBewertungen::on_tbGeruch_textChanged()
 {
     if (ui->tbGeruch->hasFocus())
         setData(ModelBewertungen::ColGeruchBemerkung, ui->tbGeruch->toPlainText());
 }
 
-void TabBewertung::setGeschmack()
+void DlgBewertungen::setGeschmack()
 {
     int value = 0;
     if (ui->rbGeschmack_0->isChecked())
@@ -579,78 +601,78 @@ void TabBewertung::setGeschmack()
     setData(ModelBewertungen::ColGeschmack, value);
 }
 
-void TabBewertung::on_rbGeschmack_0_clicked()
+void DlgBewertungen::on_rbGeschmack_0_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_1_clicked()
+void DlgBewertungen::on_rbGeschmack_1_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_2_clicked()
+void DlgBewertungen::on_rbGeschmack_2_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_3_clicked()
+void DlgBewertungen::on_rbGeschmack_3_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_4_clicked()
+void DlgBewertungen::on_rbGeschmack_4_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_5_clicked()
+void DlgBewertungen::on_rbGeschmack_5_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_6_clicked()
+void DlgBewertungen::on_rbGeschmack_6_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_7_clicked()
+void DlgBewertungen::on_rbGeschmack_7_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_8_clicked()
+void DlgBewertungen::on_rbGeschmack_8_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_9_clicked()
+void DlgBewertungen::on_rbGeschmack_9_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_10_clicked()
+void DlgBewertungen::on_rbGeschmack_10_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_11_clicked()
+void DlgBewertungen::on_rbGeschmack_11_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_rbGeschmack_12_clicked()
+void DlgBewertungen::on_rbGeschmack_12_clicked()
 {
     setGeschmack();
 }
 
-void TabBewertung::on_tbGeschmack_textChanged()
+void DlgBewertungen::on_tbGeschmack_textChanged()
 {
     if (ui->tbGeschmack->hasFocus())
         setData(ModelBewertungen::ColGeschmackBemerkung, ui->tbGeschmack->toPlainText());
 }
 
-void TabBewertung::setAntrunk()
+void DlgBewertungen::setAntrunk()
 {
     int value = 0;
     if (ui->rbAntrunk_0->isChecked())
@@ -672,58 +694,58 @@ void TabBewertung::setAntrunk()
     setData(ModelBewertungen::ColAntrunk, value);
 }
 
-void TabBewertung::on_rbAntrunk_0_clicked()
+void DlgBewertungen::on_rbAntrunk_0_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_1_clicked()
+void DlgBewertungen::on_rbAntrunk_1_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_2_clicked()
+void DlgBewertungen::on_rbAntrunk_2_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_3_clicked()
+void DlgBewertungen::on_rbAntrunk_3_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_4_clicked()
+void DlgBewertungen::on_rbAntrunk_4_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_5_clicked()
+void DlgBewertungen::on_rbAntrunk_5_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_6_clicked()
+void DlgBewertungen::on_rbAntrunk_6_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_7_clicked()
+void DlgBewertungen::on_rbAntrunk_7_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_rbAntrunk_ka_clicked()
+void DlgBewertungen::on_rbAntrunk_ka_clicked()
 {
     setAntrunk();
 }
 
-void TabBewertung::on_tbAntrunk_textChanged()
+void DlgBewertungen::on_tbAntrunk_textChanged()
 {
     if (ui->tbAntrunk->hasFocus())
         setData(ModelBewertungen::ColAntrunkBemerkung, ui->tbAntrunk->toPlainText());
 }
 
-void TabBewertung::setHaupttrunk()
+void DlgBewertungen::setHaupttrunk()
 {
     int value = 0;
     if (ui->rbHaupttrunk_0->isChecked())
@@ -739,43 +761,43 @@ void TabBewertung::setHaupttrunk()
     setData(ModelBewertungen::ColHaupttrunk, value);
 }
 
-void TabBewertung::on_rbHaupttrunk_0_clicked()
+void DlgBewertungen::on_rbHaupttrunk_0_clicked()
 {
     setHaupttrunk();
 }
 
-void TabBewertung::on_rbHaupttrunk_1_clicked()
+void DlgBewertungen::on_rbHaupttrunk_1_clicked()
 {
     setHaupttrunk();
 }
 
-void TabBewertung::on_rbHaupttrunk_2_clicked()
+void DlgBewertungen::on_rbHaupttrunk_2_clicked()
 {
     setHaupttrunk();
 }
 
-void TabBewertung::on_rbHaupttrunk_3_clicked()
+void DlgBewertungen::on_rbHaupttrunk_3_clicked()
 {
     setHaupttrunk();
 }
 
-void TabBewertung::on_rbHaupttrunk_4_clicked()
+void DlgBewertungen::on_rbHaupttrunk_4_clicked()
 {
     setHaupttrunk();
 }
 
-void TabBewertung::on_rbHaupttrunk_ka_clicked()
+void DlgBewertungen::on_rbHaupttrunk_ka_clicked()
 {
     setHaupttrunk();
 }
 
-void TabBewertung::on_tbHaupttrunk_textChanged()
+void DlgBewertungen::on_tbHaupttrunk_textChanged()
 {
     if (ui->tbHaupttrunk->hasFocus())
         setData(ModelBewertungen::ColHaupttrunkBemerkung, ui->tbHaupttrunk->toPlainText());
 }
 
-void TabBewertung::setNachtrunk()
+void DlgBewertungen::setNachtrunk()
 {
     int value = 0;
     if (ui->rbNachtrunk_0->isChecked())
@@ -799,63 +821,63 @@ void TabBewertung::setNachtrunk()
     setData(ModelBewertungen::ColNachtrunk, value);
 }
 
-void TabBewertung::on_rbNachtrunk_0_clicked()
+void DlgBewertungen::on_rbNachtrunk_0_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_1_clicked()
+void DlgBewertungen::on_rbNachtrunk_1_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_2_clicked()
+void DlgBewertungen::on_rbNachtrunk_2_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_3_clicked()
+void DlgBewertungen::on_rbNachtrunk_3_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_4_clicked()
+void DlgBewertungen::on_rbNachtrunk_4_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_5_clicked()
+void DlgBewertungen::on_rbNachtrunk_5_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_6_clicked()
+void DlgBewertungen::on_rbNachtrunk_6_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_7_clicked()
+void DlgBewertungen::on_rbNachtrunk_7_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_8_clicked()
+void DlgBewertungen::on_rbNachtrunk_8_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_rbNachtrunk_ka_clicked()
+void DlgBewertungen::on_rbNachtrunk_ka_clicked()
 {
     setNachtrunk();
 }
 
-void TabBewertung::on_tbNachtrunk_textChanged()
+void DlgBewertungen::on_tbNachtrunk_textChanged()
 {
     if (ui->tbNachtrunk->hasFocus())
         setData(ModelBewertungen::ColNachtrunkBemerkung, ui->tbNachtrunk->toPlainText());
 }
 
-void TabBewertung::setGesamteindruck()
+void DlgBewertungen::setGesamteindruck()
 {
     int value = 0;
     if (ui->rbGesamteindruck_0->isChecked())
@@ -877,53 +899,112 @@ void TabBewertung::setGesamteindruck()
     setData(ModelBewertungen::ColGesamteindruck, value);
 }
 
-void TabBewertung::on_rbGesamteindruck_0_clicked()
+void DlgBewertungen::on_rbGesamteindruck_0_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_1_clicked()
+void DlgBewertungen::on_rbGesamteindruck_1_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_2_clicked()
+void DlgBewertungen::on_rbGesamteindruck_2_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_3_clicked()
+void DlgBewertungen::on_rbGesamteindruck_3_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_4_clicked()
+void DlgBewertungen::on_rbGesamteindruck_4_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_5_clicked()
+void DlgBewertungen::on_rbGesamteindruck_5_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_6_clicked()
+void DlgBewertungen::on_rbGesamteindruck_6_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_7_clicked()
+void DlgBewertungen::on_rbGesamteindruck_7_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_rbGesamteindruck_ka_clicked()
+void DlgBewertungen::on_rbGesamteindruck_ka_clicked()
 {
     setGesamteindruck();
 }
 
-void TabBewertung::on_tbGesamteindruck_textChanged()
+void DlgBewertungen::on_tbGesamteindruck_textChanged()
 {
     if (ui->tbGesamteindruck->hasFocus())
         setData(ModelBewertungen::ColGesamteindruckBemerkung, ui->tbGesamteindruck->toPlainText());
+}
+
+void DlgBewertungen::onTableSelectionChanged(const QItemSelection &selected)
+{
+    if (selected.indexes().count())
+    {
+        const ProxyModel *proxyModel = static_cast<ProxyModel*>(ui->table->model());
+        int sudRow = proxyModel->mapRowToSource(selected.indexes()[0].row());
+        mSudId = bh->modelSud()->data(sudRow, ModelSud::ColID).toInt();
+    }
+    else
+    {
+        mSudId = -1;
+    }
+    ProxyModel *proxyModelBewertungen = static_cast<ProxyModel*>(ui->tableBewertungen->model());
+    proxyModelBewertungen->setFilterRegularExpression(QRegularExpression(QStringLiteral("^%1$").arg(mSudId)));
+    ui->tableBewertungen->selectRow(0);
+}
+
+void DlgBewertungen::onTableBewertungenSelectionChanged(const QItemSelection &selected)
+{
+    if (selected.indexes().count())
+    {
+        const ProxyModel *proxyModel = static_cast<ProxyModel*>(ui->tableBewertungen->model());
+        mRowBewertung = proxyModel->mapRowToSource(selected.indexes()[0].row());
+    }
+    else
+    {
+        mRowBewertung = -1;
+    }
+    updateValues();
+}
+
+void DlgBewertungen::onFilterChanged()
+{
+    ProxyModelSud *model = static_cast<ProxyModelSud*>(ui->table->model());
+    ProxyModelSud proxy;
+    proxy.setSourceModel(bh->modelSud());
+    proxy.setFilterStatus(ProxyModelSud::Abgefuellt | ProxyModelSud::Verbraucht);
+    ui->lblFilter->setText(QString::number(model->rowCount()) + " / " + QString::number(proxy.rowCount()));
+}
+
+void DlgBewertungen::on_tbFilter_textChanged(const QString &pattern)
+{
+    static_cast<ProxyModelSud*>(ui->table->model())->setFilterText(pattern);
+}
+
+void DlgBewertungen::on_btnAdd_clicked()
+{
+    QMap<int, QVariant> values({{ModelBewertungen::ColSudID, mSudId}});
+    int index = bh->modelBewertungen()->append(values);
+    const ProxyModel *proxyModel = static_cast<ProxyModel*>(ui->tableBewertungen->model());
+    ui->tableBewertungen->selectRow(proxyModel->mapRowFromSource(index));
+}
+
+void DlgBewertungen::on_btnRemove_clicked()
+{
+    bh->modelBewertungen()->removeRow(mRowBewertung);
+    ui->tableBewertungen->selectRow(0);
 }